List the API group for each Golang package
This lists an API group for every Golang package, based on the regular
expressions defined in config/gvk.yaml.

The goal is to deprecate the use of regexp so that new API groups can
be added manually. Using a series of regexp for this is too high a
barrier for including future APIs that might live in new Golang packages.